Add a 'delete project' option to project list
- Feature: Added a delete project icon to each project in the project list that triggers a delete confirmation flow; On approval, the selected project is deleted via the regular API.
- Bugfix: Vizier now handles empty responses gracefully
- Dev QOL: Non-returning API calls in the frontend still return a Future so that the user can know when the call succeeds/fails.
- Dev QOL: Factored the code for the project list out into its own class: ProjectListView
