+++ vizier/ui/html/index.html
@@ -5,6 +5,7 @@
     <title>Vizier</title>
     <link rel="icon" href="favicon.svg">
     <link rel="stylesheet" href="css/vizier.css"/>
+    <link rel="stylesheet" href="css/font-awesome.min.css"/>
     <!-- Include Scala.js compiled code -->
     <script type="text/javascript" src="vizier.js"></script>
     <script type="text/javascript">
